+++ runtime/cats/src/test/scala/mongo4s/cats/CatsBridgeSpec.scala
@@ -62,7 +62,7 @@ final class CatsBridgeSpec extends AsyncWordSpec, AsyncIOSpec, Matchers:
   }
 
   "cats RsBridge with strictSingleResult" should {
-    given RsBridgeConfig = RsBridgeConfig.Default.copy(strictSingleResult = true)
+    given RsBridgeConfig = RsBridgeConfig.default.withStrictSingleResult
     val strictBridge     = summon[RsBridge[IO, CatsStream[IO]]]
 
     "fail one with TooManyResults for a 2-element publisher" in {
@@ -80,7 +80,7 @@ final class CatsBridgeSpec extends AsyncWordSpec, AsyncIOSpec, Matchers:
   }
 
   "cats RsBridge with a timeout" should {
-    given RsBridgeConfig = RsBridgeConfig.Default.copy(timeout = Some(50.millis))
+    given RsBridgeConfig = RsBridgeConfig.default.withTimeout(50.millis)
     val timeoutBridge    = summon[RsBridge[IO, CatsStream[IO]]]
 
     "fail with RsBridgeError.Timeout when the publisher never completes" in {